chore: update flow for alert cmd feature
