+++ src/commands/alert/remove/processor.ts
@@ -6,13 +6,8 @@ import { getSuccessEmbed } from "ui/discord/embed"
 
 export const handler: InteractionHandler = async (msgOrInteraction) => {
   const interaction = msgOrInteraction as SelectMenuInteraction
-  const [userDiscordId, symbol, amount] = interaction.values[0].split("|")
-  const price = parseFloat(amount)
-  const { ok, log, curl } = await Defi.removeAlertPrice(
-    userDiscordId,
-    symbol,
-    price
-  )
+  const [alertId] = interaction.values[0]
+  const { ok, log, curl } = await Defi.removeAlertPrice(alertId)
 
   if (!ok) {
     throw new APIError({ description: log, curl })
